move shared code for reading/writing data blocks to abstract class
